added repeats

+++ controller.py
@@ -1,5 +1,5 @@
 from db_models import Notifies, engine, sessionmaker
-from datetime import datetime
+from datetime import datetime, timedelta
 
 Session = sessionmaker(bind=engine)
 session = Session()
@@ -14,7 +14,7 @@ class Controller:
             reply_id=data["reply"],
             time=datetime.strptime(data["time"], "%d.%m.%y:%H.%M"),
             repeats=data["repeats"],
-            chat_id=data['chat_id']
+            chat_id=data["chat_id"],
         )
         session.add(notify)
         session.commit()
@@ -29,9 +29,17 @@ class Controller:
             if datetime.strftime(i.time, "%d.%m.%y:%H.%M") == datetime.strftime(
                 datetime.utcnow(), "%d.%m.%y:%H.%M"
             ):
-                notifies.append({"text": i.message_text, "reply_id": i.reply_id,
-                    'chat_id':i.chat_id})
-                print(notifies)
-                session.delete(i)
+                notifies.append(
+                    {
+                        "text": i.message_text,
+                        "reply_id": i.reply_id,
+                        "chat_id": i.chat_id,
+                    }
+                )
+                if i.repeats != 0:
+                    i.repeats -= 1
+                    i.time = i.time+timedelta(minutes=10)
+                else:
+                    session.delete(i)
                 session.commit()
         return notifies
